% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B2}
\alias{B2}
\title{OTU data from Sinha et al. (2017)}
\format{
A dataframe containing a subset of data from Sinha et al. (2017)
}
\source{
\url{https://www.nature.com/articles/nbt.3981}
}
\usage{
B2
}
\description{
A dataset containing OTU counts reported by Sinha et al. (2017)
 in colorectal cancer patients and controls
as sequenced by sequencing laboratory B and processed by bioinformatics
laboratory 2.
}
\references{
"Sinha, Rashmi, et al. "Assessment of variation in microbial community amplicon sequencing by the Microbiome Quality Control (MBQC) project consortium." Nature biotechnology 35.11 (2017): 1077-1086.
}
\keyword{datasets}
