% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{bc_fast}
\alias{bc_fast}
\title{Bray-Curtis Distance (Fast)}
\usage{
bc_fast(p1s, p2s)
}
\arguments{
\item{p1s}{First set of proportions}

\item{p2s}{Second set of proportions}
}
\description{
Bray-Curtis Distance (Fast)
}
