% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{bray_curtis_true}
\alias{bray_curtis_true}
\title{Bray-Curtis Distance (True)}
\usage{
bray_curtis_true(p1s, p2s = NULL)
}
\arguments{
\item{p1s}{First set of proportions}

\item{p2s}{Second set of proportions, defaults to NULL}
}
\description{
Bray-Curtis Distance (True)
}
