% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{calculate_alpha}
\alias{calculate_alpha}
\title{Calculate Alpha}
\usage{
calculate_alpha(mu, my_function)
}
\arguments{
\item{mu}{Amy TODO}

\item{my_function}{Amy TODO}
}
\description{
Calculate Alpha
}
