% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_method.R
\name{delta_method}
\alias{delta_method}
\title{LNM.EM}
\usage{
delta_method(W, X = NULL, EMiter = 10, EMburn = 5, MCiter = 1000,
  MCburn = 500, stepsize = 0.01, p = 0.05, poorman = FALSE,
  interval = 0.95, my_function = c("Shannon", "Simpson"))
}
\arguments{
\item{W}{count matrix, with OTUs as columns}

\item{X}{covariate matrix}

\item{EMiter}{number of EM iterations, defaults to 10}

\item{EMburn}{number of EM iterations to burn, defaults to 5}

\item{MCiter}{number MC iterations, defaults to 1000}

\item{MCburn}{number of MC iterations to burn, defaults to 500}

\item{stepsize}{variance used for MH samples, defaults to 0.01. Tweak to adjust acceptance ratio}

\item{p}{size of purturbation used for logratios, defaults to 0.05}

\item{poorman}{boolean of whether to just use simple diagonal inverse to calculate sigma inverse. Defaults to FALSE}

\item{base}{OTU index to be used for base}
}
\description{
This function estimates the LNM model fit from Xia et al.
}
