% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivNet.R
\name{DivNet-package}
\alias{DivNet-package}
\alias{DivNet}
\title{DivNet package documentation.}
\description{
This package provides various methods and tools to analyze microbiome data
}
\details{
A complete description of all package features, along
with examples, can be found at \url{https://github.com/adw96/DivNet}.
}
