% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_diversity.R
\name{divnet}
\alias{divnet}
\title{divnet}
\usage{
divnet(W, X = NULL, fitted_model = NULL, tuning = NULL,
  perturbation = NULL, network = NULL, base = NULL, ncores = NULL,
  variance = "parametric", B = 5, nsub = NULL, ...)
}
\arguments{
\item{variance}{method to get variance of estimates. Current options are "parametric" for parametric bootstrap, "nonparametric" for nonparametric bootstrap, and "none" for no variance estimates}

\item{fitted_aitchison}{object produced by fit_aitchison}
}
\description{
divnet
}
\author{
Amy Willis
}
