% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmmBars.R
\name{dmmBarsT}
\alias{dmmBarsT}
\title{dmmBarsT}
\usage{
dmmBarsT(out, W, main = "DMM Model Fit")
}
\arguments{
\item{out}{model fit from package DirichletMultinomial}

\item{W}{raw count matrix used in model fit}

\item{main}{desired title for plot, defaults to DMM Model Fit}
}
\description{
Function to plot DMM model fit, with empirical 95% bars, transposed
}
