% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_diversity.R
\name{estimate_diversity}
\alias{estimate_diversity}
\title{estimate_diversity}
\usage{
estimate_diversity(fitted_aitchison)
}
\arguments{
\item{fitted_aitchison}{object produced by fit_aitchison}
}
\description{
estimate_diversity
}
