% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{euc_fast}
\alias{euc_fast}
\title{Euclidean Distance (Fast)}
\usage{
euc_fast(p1s, p2s)
}
\arguments{
\item{p1s}{First set of proportions}

\item{p2s}{Second set of proportions}
}
\description{
Euclidean Distance (Fast)
}
