% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{euclidean_true}
\alias{euclidean_true}
\title{Euclidean Distance (True)}
\usage{
euclidean_true(p1s, p2s = NULL)
}
\arguments{
\item{p1s}{First set of proportions}

\item{p2s}{Second set of proportions, defaults to NULL}
}
\description{
Euclidean Distance (True)
}
