% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOTU.R
\name{extractOTU}
\alias{extractOTU}
\title{extractOTU}
\usage{
extractOTU(W, nrand, nmost)
}
\arguments{
\item{W}{raw count matrix, with OTUs as columns}

\item{nrand}{number of random OTUs desired}

\item{nmost}{number of most prevalent OTUs desired}
}
\description{
This function extracts the desired combination of random and most prevalent OTUs, including 0.
}
