% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_aitchison.R
\name{fit_aitchison}
\alias{fit_aitchison}
\title{fit_aitchison}
\usage{
fit_aitchison(
  W,
  X = NULL,
  tuning = NULL,
  perturbation = NULL,
  network = NULL,
  base = NULL,
  ncores = NULL,
  ...
)
}
\arguments{
\item{W}{count matrix, with columns listing the OTUs and rows listing the samples}

\item{X}{covariate matrix (optional)}

\item{tuning}{settings for tuning the MC-MH algorithm. Options include NULL (defaults to "fast"), "fast", "careful" or a named list with components EMiter (number of EM iterations; 6 for fast, 10 for careful), EMburn (number of EM iterations to burn; 3 for fast, 5 for careful), MCiter (number of MC iterations; 500 for fast, 1000 for careful), MCburn (number of MC iterations to burn; 250 for fast, 500 for careful) and stepsize (variance used for MH samples; 0.01 for both fast and careful)}

\item{perturbation}{size of purturbation for zero counts, defaults to NULL (0.05)}

\item{network}{How to estimate network. Defaults to NULL (the default), "default" (generalised inverse, aka naive). Other options include "diagonal", "stars" (requires glasso and SpiecEasi to be installed), or a function that you want to use to estimate the network}

\item{base}{OTU index to be used for base. if NULL, will use most common taxon}

\item{ncores}{number of cores to use for MH sampling, defaults to 1}

\item{...}{additional arguments to be supplied to the network function}
}
\description{
This function estimates the parameters of the log-ratio model
}
\author{
Bryan Martin

Amy Willis
}
