% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getPurt}
\alias{getPurt}
\title{getPurt}
\usage{
getPurt(Wi, base, perturbation = 0.05)
}
\arguments{
\item{Wi}{raw count data row}

\item{base}{base OTU value}

\item{p}{how much to purturb zero counts, defaults to 0.05}
}
\description{
This function purturbates a row of the raw count matrix. Likely will not be used directly.
}
