% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lee}
\alias{Lee}
\title{phyloseq object from Lee et al (2015)}
\format{
A phyloseq object containing the data from Lee et al (2015)
}
\source{
\url{https://astrobiomike.github.io/amplicon/workflow_ex}
}
\usage{
Lee
}
\description{
A dataset containing the abundance table, covariate information, and taxonomy of a 16S study of basalts on the Dorado Outcrop
}
\references{
"Microbial Communities on Seafloor Basalts at Dorado Outcrop Reflect Level of Alteration and Highlight Global Lithic Clades" by Lee et. al (2015), Frontiers in Microbiology
}
\keyword{datasets}
