% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divnet_alpha_estimates.R
\name{make_alpha_estimates}
\alias{make_alpha_estimates}
\title{Make alpha estimates}
\usage{
make_alpha_estimates(dn)
}
\arguments{
\item{dn}{A DivNet object.}
}
\value{
An object of class \code{alpha_estimate} containing alpha estimate
information in \code{dn}. 

export
}
\description{
This function is a wrapper for \code{alpha_estimates} in \code{breakaway}.
}
