% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{make_composition}
\alias{make_composition}
\title{make_composition}
\usage{
make_composition(W)
}
\arguments{
\item{W}{raw count matrix, with OTUs as columns}
}
\description{
This function calculates the observed composition from raw counts
}
