% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_w}
\alias{make_w}
\title{Simulate counts}
\usage{
make_w(mu, Sigma, mm, base = NULL)
}
\arguments{
\item{mu}{Mean matrix. Rows are samples}

\item{Sigma}{Covariance matrix}

\item{mm}{Sampling depth (= number of individuals observed)}

\item{base}{The base taxon for logratio transformations}
}
\description{
Simulate counts
}
