% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeComp.R
\name{makeComp}
\alias{makeComp}
\title{makeComp}
\usage{
makeComp(W)
}
\arguments{
\item{W}{raw count matrix, with OTUs as columns}
}
\description{
This function calculates the observed composition from raw counts
}
