% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCrow.R
\name{MCrow}
\alias{MCrow}
\title{MCrow}
\usage{
MCrow(Yi, Wi, eYi, Q, base, sigInv, MCiter, stepsize = 1)
}
\arguments{
\item{Yi}{row of logratio matrix}

\item{Wi}{corresponding row of count matrix}

\item{eYi}{current expected value of logratio matrix}

\item{Q}{number of OTUs minus the base, or length of Yi}

\item{base}{OTU index used for base}

\item{sigInv}{current estimate of sigma inverse}

\item{MCiter}{number of MC samples to generate}

\item{stepsize}{variance used for MH samples, defaults to 1. Tweak to adjust acceptance ratio}
}
\description{
This function simulates MC step for a single row. Should not need to be used by user directly.
}
