% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcrow_MCrow}
\alias{mcrow_MCrow}
\title{MCrow}
\usage{
mcrow_MCrow(Yi, Wi, eYi, base, sigInv, MCiter, stepsize)
}
\arguments{
\item{Yi}{row of logratio matrix}

\item{Wi}{corresponding row of count matrix}

\item{eYi}{current expected value of logratio matrix}

\item{base}{OTU index used for base (1-based index)}

\item{sigInv}{current estimate of sigma inverse}

\item{MCiter}{number of MC samples to generate}

\item{stepsize}{variance used for MH samples. Tweak to adjust acceptance ratio}
}
\description{
This function simulates MC step for a single row.
}
\author{
Ryan Moore
}
