% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicrobiomePack.R
\name{MicrobiomePack-package}
\alias{MicrobiomePack-package}
\alias{MicrobiomePack}
\title{MicrobiomePack package documentation.}
\description{
This package provides various methods and tools to analyze microbiome data
}
\details{
A complete description of all package features, along
with examples, can be found at \url{https://github.com/martinbryan/MicrobiomePack}.
}
