% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betta_posthoc.R
\name{new_betta_posthoc}
\alias{new_betta_posthoc}
\title{Create a betta_posthoc object with proper class}
\usage{
new_betta_posthoc(results_df, cld_df)
}
\arguments{
\item{results_df}{Pairwise comparison results}

\item{cld_df}{Compact letter display results}
}
\value{
A betta_posthoc object
}
\description{
Create a betta_posthoc object with proper class
}
\keyword{internal}
