% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylodivnet.R
\name{phylodivnet}
\alias{phylodivnet}
\title{phylodivnet}
\usage{
phylodivnet(W, X = NULL, trees = NULL, tuning = NULL,
  perturbation = NULL, network = NULL, base = NULL, ncores = NULL,
  B = 5, alpha = 0.05, ...)
}
\arguments{
\item{W}{An abundance table with taxa as columns and samples as rows; or a phyloseq object}

\item{X}{The covariate matrix, with samples as rows and variables as columns. Note that the estimated UniFrac distance of biological replicates will be zero.}

\item{trees}{A vector of file paths to files containing the tree estimates in Newick format.}

\item{tuning}{settings for tuning the MC-MH algorithm. Options include NULL (defaults to "fast"), "fast", "careful" or a named list with components EMiter (number of EM iterations; 6 for fast, 10 for careful), EMburn (number of EM iterations to burn; 3 for fast, 5 for careful), MCiter (number of MC iterations; 500 for fast, 1000 for careful), MCburn (number of MC iterations to burn; 250 for fast, 500 for careful) and stepsize (variance used for MH samples; 0.01 for both fast and careful)}

\item{perturbation}{Perturbation magnitude for zero values when calculating logratios.}

\item{network}{How to estimate network. Defaults to NULL (the default), "default" (generalised inverse, aka naive). Other options include "diagonal", "stars" (requires glasso and SpiecEasi to be installed), or a function that you want to use to estimate the network}

\item{base}{Base taxon index.}

\item{ncores}{Number of cores to use for parallelization}

\item{B}{Number of bootstrap iterations for estimating the variance.}

\item{alpha}{For a 100*(1-alpha) percent confidence interval. Defaults to 0.05.}

\item{...}{Additional parameters to be passed to the network function}
}
\description{
phylodivnet
}
\author{
Amy Willis
}
