% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pick_base}
\alias{pick_base}
\title{pick_base}
\usage{
pick_base(W)
}
\arguments{
\item{W}{A taxon abundance matrix (taxa as columns)}
}
\description{
Picks the base taxon to be used in divnet fit. If no taxon is detected
in all samples, returns error; in this case, we recommend manually choosing 
a few taxa fairly abundant across samples and for each such taxon
fitting divnet specifying this taxon as the base taxon.
}
\author{
Amy Willis
}
