% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3functions.R
\name{plot.diversityEstimates}
\alias{plot.diversityEstimates}
\title{Plot function}
\usage{
\method{plot}{diversityEstimates}(x, ...)
}
\arguments{
\item{x}{An object of class diversityEstimates}

\item{...}{other arguments to be passed to plot}
}
\value{
An object of class ggplot
}
\description{
TODO make more like the phyloseq plot richness
}
