% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3functions.R
\name{print.diversityEstimates}
\alias{print.diversityEstimates}
\title{Print function}
\usage{
\method{print}{diversityEstimates}(x, ...)
}
\arguments{
\item{x}{An object of class diversityEstimates}

\item{...}{other arguments to be passed to print}
}
\description{
Print function
}
