% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mc}
\alias{run_mc}
\title{run_mc}
\usage{
run_mc(logratios, counts, expected_logratios, sigma_inverse, base_otu,
  mc_iters, iters_to_burn, stepsize)
}
\arguments{
\item{logratios}{The logratio matrix.  Dim: nsamples x (notus - 1).}

\item{counts}{The corresponding count matrix.  Dim: nsamples x notus.}

\item{expected_logratios}{The current expected value of the logratio matrix.  Dim: nsamples x (notus - 1).}

\item{sigma_inverse}{The current estimate of sigma inverse.}

\item{base_otu}{The index of the otu to remove from the input vector}

\item{mc_iters}{Total number of mc iters to run}

\item{iters_to_burn}{The number of MC iters to burn}

\item{stepsize}{Variance used for MH samples}
}
\description{
Simulate MC step.
}
\author{
Ryan Moore
}
