% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleMult.R
\name{simpleMult}
\alias{simpleMult}
\title{simpleMult}
\usage{
simpleMult(W, main = "Simple Multinomial Variance")
}
\arguments{
\item{W}{observed count matrix}

\item{main}{desired title for plot, defaults to Simple Multinomial Variance}
}
\description{
Function to plot simple multimial model fit, with error bars
}
