% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmat.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(sigma, W, base, perturbation, ncores, ...)
}
\arguments{
\item{sigma}{current estimate of sigma}

\item{W}{corresponding count matrix}

\item{base}{OTU index used for base}

\item{perturbation}{size of purturbation used for to_log_ratios, defaults to 0.05}

\item{ncores}{number of cores to use, defaults to 1}

\item{...}{other arguments to pass}
}
\description{
Estimate the network using the package pulsar to select the tuning parameter.
Type `DivNet::stars` to see the chosen defaults. To modify the defaults,
create a new function and pass it as the `network` argument to MCmat.
}
