% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3functions.R
\name{testDiversity}
\alias{testDiversity}
\title{Test diversity}
\usage{
testDiversity(dv, h0 = "shannon")
}
\arguments{
\item{dv}{An object of class diversityEstimates. The variable `X` used for the construction}

\item{h0}{The alpha-diversity index to be tested for equality}
}
\value{
A data frame similar to the output of `lm`
}
\description{
Hypothesis testing for alpha-diversity.
}
\references{
Willis, A., Bunge, J., and Whitman, T. (2017). Improved detection of changes in species richness in high diversity microbial communities. \emph{JRSS-C.}
}
