% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{to_composition}
\alias{to_composition}
\title{to_composition}
\usage{
to_composition(Y, base = NULL)
}
\arguments{
\item{Y}{A vector of logratios}

\item{base}{base taxon used to calculate logratios}
}
\description{
This function transforms a vector of logratios Y to a composition matrix X
}
