% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{to_log_ratios}
\alias{to_log_ratios}
\title{to_log_ratios}
\usage{
to_log_ratios(W, base, perturbation = 0.05)
}
\arguments{
\item{W}{Any compositional (count or proportion) data to be transformed, with OTUs as columns}

\item{base}{base OTU value}

\item{perturbation}{how much to purturb zero counts, defaults to 0.05}
}
\description{
This function transforms from original scale to logratios
}
