% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{toComposition}
\alias{toComposition}
\title{toComposition}
\usage{
toComposition(Y, base = NULL)
}
\arguments{
\item{Y}{A vector of logratios}

\item{base}{base taxon used to calculate logratios}
}
\description{
This function transforms a vector of logratios Y to a composition matrix X
}
