% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{toCompositionMatrix}
\alias{toCompositionMatrix}
\title{toCompositionMatrix}
\usage{
toCompositionMatrix(Y, base = NULL)
}
\arguments{
\item{Y}{A matrix of logratios. The samples are listed across the rows and the taxa are listed across the columns}

\item{base}{base taxon used to calculate logratios}
}
\description{
This function transforms a logratio matrix Y to a composition matrix X
}
