% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichpart.R
\name{whichpart}
\alias{whichpart}
\title{whichpart}
\usage{
whichpart(x, n = 75)
}
\arguments{
\item{x}{OTU data}

\item{n}{number of OTUs to extract, defaults to 75}
}
\description{
This function selects the n most prevalent OTUs
}
