% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wsim.R
\name{Wsim}
\alias{Wsim}
\title{Wsim}
\usage{
Wsim(out, W, X = NULL, niter = 1000, fast = FALSE)
}
\arguments{
\item{out}{model fit from LNM.EM}

\item{W}{raw count matrix used in model fit}

\item{niter}{number of simulations, defaults to 1000}

\item{fast}{whether or not to simulate from diagonal sigma with rmvn}
}
\description{
Function to simulate raw counts W from model fit
}
