% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YtoW.R
\name{YtoW}
\alias{YtoW}
\title{YtoW}
\usage{
YtoW(Y, M, base)
}
\arguments{
\item{Y}{logratio matrix}

\item{M}{vector of observed counts by row}

\item{base}{base value used to calculate logratios}
}
\description{
This function transforms logratio matrix Y to counts W
}
