% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YtoX.R
\name{YtoX}
\alias{YtoX}
\title{YtoX}
\usage{
YtoX(Y, base)
}
\arguments{
\item{Y}{logratio matrix}

\item{base}{base value used to calculate logratios}
}
\description{
This function transforms logratio matrix Y to composition X
}
