% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_diversity.R
\name{divnet}
\alias{divnet}
\title{divnet}
\usage{
divnet(W, X = NULL, fitted_model = NULL, tuning = NULL,
  perturbation = NULL, network = NULL, base = NULL, ncores = NULL,
  variance = "parametric", B = 5, nsub = NULL, ...)
}
\arguments{
\item{W}{An abundance table with taxa as columns and samples as rows; or a phyloseq object}

\item{X}{The covariate matrix, with samples as rows and variables as columns. Defaults to NULL (sample_names are the covariates).}

\item{fitted_model}{object produced by fit_aitchison. Defaults to NULL.}

\item{tuning}{settings for tuning the MC-MH algorithm. Options include NULL (defaults to "fast"), "fast", "careful" or a named list with components EMiter (number of EM iterations; 6 for fast, 10 for careful), EMburn (number of EM iterations to burn; 3 for fast, 5 for careful), MCiter (number of MC iterations; 500 for fast, 1000 for careful), MCburn (number of MC iterations to burn; 250 for fast, 500 for careful) and stepsize (variance used for MH samples; 0.01 for both fast and careful)}

\item{perturbation}{Perturbation magnitude for zero values when calculating logratios.}

\item{network}{How to estimate network. Defaults to NULL (the default), "default" (generalised inverse, aka naive). Other options include "diagonal", "stars" (requires glasso and SpiecEasi to be installed), or a function that you want to use to estimate the network}

\item{base}{The column index of the base taxon in the columns of W, or the name of the taxon (must be a column name of W, or a taxon name if W is a phyloseq object). If NULL, will use `pick_base` to choose a taxon. If no taxa are observed in all samples, an error will be thrown. In that case, we recommend trying a number of different highly abundant taxa to confirm the results are robust to the taxon choice.}

\item{ncores}{Number of cores to use for parallelization}

\item{variance}{method to get variance of estimates. Current options are "parametric" for parametric bootstrap, "nonparametric" for nonparametric bootstrap, and "none" for no variance estimates}

\item{B}{Number of bootstrap iterations for estimating the variance.}

\item{nsub}{Number of subsamples for nonparametric bootstrap. Defaults to half the number of observed samples.}

\item{...}{Additional parameters to be passed to the network function}
}
\description{
divnet
}
\author{
Amy Willis
}
