% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_aitchison.R
\name{fit_aitchison}
\alias{fit_aitchison}
\title{LNM.EM}
\usage{
fit_aitchison(W, X = NULL, EMiter = 10, EMburn = 5, MCiter = 1000,
  MCburn = 500, stepsize = 0.01, perturbation = 0.05,
  network = "default", base = NULL, ncores = 1, ...)
}
\arguments{
\item{W}{count matrix, with OTUs as columns}

\item{X}{covariate matrix (optional)}

\item{EMiter}{number of EM iterations, defaults to 10}

\item{EMburn}{number of EM iterations to burn, defaults to 5}

\item{MCiter}{number MC iterations, defaults to 1000}

\item{MCburn}{number of MC iterations to burn, defaults to 500}

\item{stepsize}{variance used for MH samples, defaults to 0.01. Tweak to adjust acceptance ratio}

\item{perturbation}{size of purturbation used for toLogRatios, defaults to 0.05}

\item{network}{How to estimate network. Defaults to "default" (generalised inverse, aka naive). Other options include "diagonal", "stars" (requires glasso and SpiecEasi to be installed), or a function that you want to use to estimate the network}

\item{base}{OTU index to be used for base. if NULL, will use most common taxon}

\item{ncores}{number of cores to use, defaults to 1}

\item{...}{additional arguments to be supplied to the network function}
}
\description{
This function estimates the LNM model fit from Xia et al.
}
\author{
Bryan Martin

Amy Willis
}
