% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmat.R
\name{MCmat}
\alias{MCmat}
\title{MCmat}
\usage{
MCmat(Y, W, eY, N, Q, base, sigma, MCiter, stepsize = 1,
  perturbation = 0.05, network = "default", ncores = 1, ...)
}
\arguments{
\item{Y}{logratio matrix}

\item{W}{corresponding count matrix}

\item{eY}{current expected value of logratio matrix}

\item{N}{number of samples, or nrow of Y}

\item{Q}{number of OTUs minus the base, or ncol of Y}

\item{base}{OTU index used for base}

\item{sigma}{current estimate of sigma}

\item{MCiter}{number of MC samples to generate}

\item{stepsize}{variance used for MH samples, defaults to 1. Tweak to adjust acceptance ratio}

\item{perturbation}{size of purturbation used for to_log_ratios, defaults to 0.05}

\item{network}{How to estimate network. Defaults to "default" (generalised inverse, aka naive). Other options include "diagonal", or a function that takes a sample covariance matrix and returns an estimate of the inverse covariance matrix (eg glasso or SpiecEasi)}

\item{ncores}{number of cores to use, defaults to 1}

\item{...}{additional arguments to be supplied to the network function}
}
\description{
This function simulates MC step for an entire matrix. Should not need to be used by user directly; available to help with determining network estimation.
}
\author{
Bryan Martin

Amy Willis
}
