% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_diversity.R
\name{plot.diversityEstimates}
\alias{plot.diversityEstimates}
\title{Plot function}
\usage{
\method{plot}{diversityEstimates}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
TODO make more like the phyloseq plot richness
}
