% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_diversity.R
\name{print.diversityEstimates}
\alias{print.diversityEstimates}
\title{Print function}
\usage{
\method{print}{diversityEstimates}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
Print function
}
