% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{stars}
\alias{stars}
\title{stars}
\usage{
stars(sigma, W, base, perturbation, ncores, ...)
}
\arguments{
\item{sigma}{current estimate of sigma}

\item{W}{corresponding count matrix}

\item{base}{OTU index used for base}

\item{perturbation}{size of purturbation used for to_log_ratios, defaults to 0.05}

\item{ncores}{number of cores to use, defaults to 1}

\item{...}{other arguments to pass

Estimate the network using the package SpiecEasi}
}
\description{
stars
}
